+++ CHANGELOG.md
@@ -10,6 +10,14 @@ Of note:
 - The changelog 2015.5.2 has been rewritten from each commit content.
 - This file may be amended entirely in the future to adhere to the [GNU Changelog style](https://www.gnu.org/prep/standards/html_node/Style-of-Change-Logs.html#Style-of-Change-Logs)
 
+## [2015.12.1]
+### Added
+- Solved [exercice for 2015, day 12](src/year_2015/day_12.rs).
+
+## [2015.11.1]
+### Added
+- Solved [exercice for 2015, day 11](src/year_2015/day_11.rs).
+
 ## [2015.10.1]
 ### Added
 - Solved [exercice for 2015, day 10](src/year_2015/day_10.rs).